// lib/types.ts
export interface NewsItem {
  id: string
  title: string
  content: string
  date: string
  author: string
}

export interface TeamMember {
  id: string
  name: string
  role: string
  image: string
  link: string
}

export interface SubServer {
  id: string
  name: string
  description: string
  image: string
  players: string
}